% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{This is the first function.}
\usage{
foo(args = NA)
}
\arguments{
\item{args}{= NA you need no argument}
}
\value{
several strings and the number of cores in
         your computer.
}
\description{
\code{foo} return some info
}
\examples{
foo (args = NA)
}

